+++ kona-pkix/src/test/resources/gen_certs.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 #
-# Copyright (C) 2022, 2023, THL A29 Limited, a Tencent company. All rights reserved.
+# Copyright (C) 2022, 2024, THL A29 Limited, a Tencent company. All rights reserved.
 # DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
 #
 # This code is free software; you can redistribute it and/or modify it
@@ -41,7 +41,7 @@ EOF
 
 echo password > password
 
-OPENSSL=babassl
+OPENSSL=tongsuo
 
 ##### CAs
 echo "RSA key CA, signed by SHA256withRSA"
