+++ kona-ssl/src/test/java/com/tencent/kona/ssl/interop/OpenSSL.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022, 2023, THL A29 Limited, a Tencent company. All rights reserved.
+ * Copyright (C) 2022, 2024, THL A29 Limited, a Tencent company. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -29,7 +29,7 @@ public class OpenSSL extends AbstractProduct {
 
     public static final OpenSSL DEFAULT = new OpenSSL(
             "OpenSSL",
-            System.getProperty("test.openssl.path", "babassl"));
+            System.getProperty("test.openssl.path", "tongsuo"));
 
     private final String name;
     private final Path path;
