+++ kona-ssl/src/test/java/com/tencent/kona/ssl/interop/TongsuoClient.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022, 2023, THL A29 Limited, a Tencent company. All rights reserved.
+ * Copyright (C) 2022, 2024, THL A29 Limited, a Tencent company. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -19,22 +19,17 @@
 
 package com.tencent.kona.ssl.interop;
 
-import com.tencent.kona.ssl.TestUtils;
-
 import java.io.IOException;
 import java.util.Collections;
 
 /*
- * The OpenSSL/BabaSSL server.
+ * The OpenSSL/Tongsuo client.
  */
-public class BabaSSLServer extends AbstractServer {
+public class TongsuoClient extends AbstractClient {
 
-    private final int port;
     private final String command;
 
-    private Process process;
-
-    public BabaSSLServer(Builder builder) throws IOException {
+    public TongsuoClient(Builder builder) {
         CertTuple certTuple = builder.getCertTuple();
         FileCert caCert = (FileCert) certTuple.trustedCert();
         FileCert signEeCert = (FileCert) certTuple.signEndEntityCert();
@@ -45,69 +40,82 @@ public class BabaSSLServer extends AbstractServer {
 
         String serverName = builder.getServerName();
         String alpn = Utilities.join(builder.getAppProtocols());
-
-        port = builder.getPort() == 0
-                ? TestUtils.getFreePort()
-                : builder.getPort();
+        String sessIn = builder.sessIn();
+        String sessOut = builder.sessOut();
 
         command = Utilities.join(" ",
+                "echo " + builder.getMessage(),
+                // Wait for the server response, e.g. session ticket.
+                "|", "sleep 1", "|",
                 getProduct().getPath().toString(),
-                "s_server",
-                Utilities.DEBUG ? "-state -debug -trace" : "",
+                "s_client",
+                Utilities.DEBUG ? "-state -debug -trace" : "-quiet",
                 "-enable_ntls", "-ntls",
                 Utilities.joinOptValue("-CAfile",
                         caCert == null ? null : caCert.certPath()),
                 "-sign_cert " + signEeCert.certPath(),
-                "-sign_key " + signEeCert.keyPath(),
+                "-sign_key " +  signEeCert.keyPath(),
                 "-enc_cert " + encEeCert.certPath(),
-                "-enc_key " + encEeCert.keyPath(),
-                builder.getClientAuth() == ClientAuth.REQUIRED ? "-Verify 2" : "",
-                builder.isUseSessTicket() ? "" : "-no_ticket",
+                "-enc_key " +  encEeCert.keyPath(),
+                Utilities.joinOptValue("-cipher",
+                        TongsuoUtils.cipherSuite(builder.getCipherSuite())),
                 Utilities.joinOptValue("-servername", serverName),
                 Utilities.joinOptValue("-alpn", alpn),
-                "-WWW",
-                "-accept " + port);
+                builder.isUseSessTicket() ? "" : "-no_ticket",
+                Utilities.joinOptValue("-sess_in", sessIn),
+                Utilities.joinOptValue("-sess_out", sessOut),
+                "-no_ign_eof");
     }
 
-    public static class Builder extends AbstractServer.Builder {
+    public static class Builder extends AbstractClient.Builder {
+
+        private String sessIn;
+        private String sessOut;
+
+        public String sessIn() {
+            return sessIn;
+        }
+
+        public Builder sessIn(String sessIn) {
+            this.sessIn = sessIn;
+            return this;
+        }
+
+        public String sessOut() {
+            return sessOut;
+        }
+
+        public Builder sessOut(String sessOut) {
+            this.sessOut = sessOut;
+            return this;
+        }
 
         @Override
-        public BabaSSLServer build() throws IOException {
-            return new BabaSSLServer(this);
+        public TongsuoClient build() {
+            return new TongsuoClient(this);
         }
     }
 
     @Override
     public Product getProduct() {
-        return BabaSSL.DEFAULT;
+        return Tongsuo.DEFAULT;
     }
 
     @Override
-    public int getPort() {
-        System.out.println("Waiting for port...");
-        if (!Utilities.waitFor(BabaSSLServer::isAlive, this)) {
-            throw new RuntimeException("Server doesn't start in time.");
-        }
-
-        return port;
-    }
-
-    @Override
-    public boolean isAlive() {
-        return process != null && process.isAlive();
-    }
-
-    @Override
-    public void accept() throws IOException {
-        process = ProcUtils.shellProc(command, getLogPath(), Collections.emptyMap());
+    public void connect(String host, int port) throws IOException {
+        String server = host + ":" + port;
+        Process process = ProcUtils.shellProc(
+                String.join(" ", command, "-connect", server),
+                getLogPath(),
+                Collections.emptyMap());
         try {
             process.waitFor();
         } catch (InterruptedException e) {
-            throw new RuntimeException("Server is interrupted!", e);
+            throw new RuntimeException("Client is interrupted!", e);
         }
 
         if (process.exitValue() != 0) {
-            throw new SSLTestException("Server exited abnormally!");
+            throw new SSLTestException("Client exited abnormally!");
         }
     }
 
@@ -115,9 +123,5 @@ public class BabaSSLServer extends AbstractServer {
     public void close() throws IOException {
         printLog();
         deleteLog();
-
-        if (isAlive()) {
-            Utilities.destroyProcess(process);
-        }
     }
 }
