+++ kona-ssl/src/test/java/com/tencent/kona/ssl/tls/JdkServerTongsuoClientTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022, 2023, THL A29 Limited, a Tencent company. All rights reserved.
+ * Copyright (C) 2022, 2024, THL A29 Limited, a Tencent company. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -25,11 +25,10 @@ import com.tencent.kona.ssl.interop.Client;
 import com.tencent.kona.ssl.interop.ClientAuth;
 import com.tencent.kona.ssl.interop.FileCert;
 import com.tencent.kona.ssl.interop.HashAlgorithm;
-import com.tencent.kona.ssl.interop.JdkClient;
-import com.tencent.kona.ssl.interop.JdkProcClient;
+import com.tencent.kona.ssl.interop.JdkServer;
 import com.tencent.kona.ssl.interop.KeyAlgorithm;
 import com.tencent.kona.ssl.interop.NamedGroup;
-import com.tencent.kona.ssl.interop.OpenSSLServer;
+import com.tencent.kona.ssl.interop.OpenSSLClient;
 import com.tencent.kona.ssl.interop.Protocol;
 import com.tencent.kona.ssl.interop.Server;
 import com.tencent.kona.ssl.interop.ServerCaller;
@@ -42,19 +41,17 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
-import javax.net.ssl.SSLContext;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.nio.file.StandardOpenOption;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
 /**
- * The interop testing between BabaSSL(OpenSSL) server and JDK client.
+ * The interop testing between JDK server and OpenSSL client.
  */
-public class BabaSSLServerJdkClientTest {
+public class JdkServerTongsuoClientTest {
 
     private static final FileCert ECDSA_INTCA_CERT = new FileCert(
             KeyAlgorithm.EC, SignatureAlgorithm.ECDSA, HashAlgorithm.SHA256,
@@ -74,28 +71,21 @@ public class BabaSSLServerJdkClientTest {
             "ee-sm2sm2-sm2sm2-sm2sm2.crt",
             "ee-sm2sm2-sm2sm2-sm2sm2.key");
 
-    private static final Path PAGE_FILE = Paths.get("build", "page");
+    private static final Path SESS_FILE = Paths.get("build", "openssl.sess");
 
     @BeforeAll
     public static void setup() throws IOException {
+        deleteSessFile();
         TestUtils.addProviders();
-        deleteWebPage();
-        createWebPage();
-    }
-
-    private static void createWebPage() throws IOException {
-        Files.write(PAGE_FILE,
-                "OpenSSL server".getBytes(Utilities.CHARSET),
-                StandardOpenOption.CREATE);
     }
 
     @AfterAll
     public static void clean() throws IOException {
-        deleteWebPage();
+        deleteSessFile();
     }
 
-    private static void deleteWebPage() throws IOException {
-        Files.deleteIfExists(PAGE_FILE);
+    private static void deleteSessFile() throws IOException {
+        Files.deleteIfExists(SESS_FILE);
     }
 
     @Test
@@ -233,19 +223,45 @@ public class BabaSSLServerJdkClientTest {
                 clientAuth);
     }
 
+    @Test
+    public void testCertSelectionWithSignatureSchemeOnTLS13()
+            throws Exception {
+        // The JDK server selects ECDSA_EE_CERT due to
+        // the OpenSSL client prefers to ECDSA_SECP256R1_SHA256.
+        connect(
+                new FileCert[] { ECDSA_INTCA_CERT },
+                new FileCert[] { SM_EE_CERT, ECDSA_EE_CERT },
+                Protocol.TLSV1_3,
+                CipherSuite.TLS_AES_128_GCM_SHA256,
+                NamedGroup.SECP256R1,
+                SignatureScheme.ECDSA_SECP256R1_SHA256,
+                ClientAuth.NONE);
+
+        // The JDK server selects SM_EE_CERT due to
+        // the OpenSSL client prefers to SM2SIG_SM3.
+        connect(
+                new FileCert[] { SM_INTCA_CERT },
+                new FileCert[] { ECDSA_EE_CERT, SM_EE_CERT },
+                Protocol.TLSV1_3,
+                CipherSuite.TLS_AES_128_GCM_SHA256,
+                NamedGroup.SECP256R1,
+                SignatureScheme.SM2SIG_SM3,
+                ClientAuth.NONE);
+    }
+
     private void connect(
             FileCert[] trustedCerts,
             FileCert[] eeCerts,
             Protocol clientProtocol,
             CipherSuite clientCipherSuite,
             NamedGroup clientNamedGroup,
-            SignatureScheme signatureScheme,
+            SignatureScheme clientSignatureScheme,
             ClientAuth clientAuth) throws Exception {
         CertTuple certTuple = new CertTuple(trustedCerts, eeCerts);
 
         ExecutorService executor = Executors.newFixedThreadPool(1);
 
-        OpenSSLServer.Builder serverBuilder = new OpenSSLServer.Builder();
+        JdkServer.Builder serverBuilder = new JdkServer.Builder();
         serverBuilder.setCertTuple(certTuple);
         serverBuilder.setProtocols(Protocol.TLSV1_3, Protocol.TLSV1_2);
         serverBuilder.setCipherSuites(
@@ -259,32 +275,15 @@ public class BabaSSLServerJdkClientTest {
             executor.submit(new ServerCaller(server));
             Utilities.waitFor(Server::isAlive, server);
 
-            try (Client client = createProcClient(
+            try (Client client = createClient(
                     certTuple, clientProtocol,
                     clientCipherSuite, clientNamedGroup,
-                    signatureScheme)) {
+                    clientSignatureScheme)) {
                 client.connect("127.0.0.1", server.getPort());
             }
         } finally {
             executor.shutdown();
         }
-    }
-
-    private JdkProcClient createProcClient(
-            CertTuple certTuple, Protocol protocol,
-            CipherSuite cipherSuite, NamedGroup namedGroup,
-            SignatureScheme signatureScheme) throws Exception {
-        JdkProcClient.Builder builder = new JdkProcClient.Builder();
-        builder.setCertTuple(certTuple);
-        builder.setProtocols(protocol);
-        builder.setCipherSuites(cipherSuite);
-        builder.setNamedGroups(namedGroup);
-        builder.setSignatureSchemes(signatureScheme);
-        builder.setMessage(
-                // An HTTP request asks to access the page.
-                String.format("GET /%s HTTP/1.1\r\n", PAGE_FILE));
-        builder.setReadResponse(true);
-        return builder.build();
     }
 
     @Test
@@ -389,7 +388,7 @@ public class BabaSSLServerJdkClientTest {
 
         ExecutorService executor = Executors.newFixedThreadPool(1);
 
-        OpenSSLServer.Builder serverBuilder = new OpenSSLServer.Builder();
+        JdkServer.Builder serverBuilder = new JdkServer.Builder();
         serverBuilder.setCertTuple(certTuple);
         serverBuilder.setProtocols(Protocol.TLSV1_3, Protocol.TLSV1_2);
         serverBuilder.setCipherSuites(
@@ -397,33 +396,30 @@ public class BabaSSLServerJdkClientTest {
                 CipherSuite.TLS_SM4_GCM_SM3,
                 CipherSuite.TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA256,
                 CipherSuite.TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256);
-        serverBuilder.setUseSessTicket(isUseSessTicket);
         serverBuilder.setClientAuth(clientAuth);
 
-        try (Server server = serverBuilder.build()) {
+        try (JdkServer server = serverBuilder.build()) {
             executor.submit(new ServerCaller(server));
             Utilities.waitFor(Server::isAlive, server);
 
-            SSLContext context = null;
             long firstCreationTime = 0;
-            try (JdkClient client = createClient(
-                    certTuple, clientProtocol,
-                    clientCipherSuite, clientNamedGroup,
-                    clientSignatureScheme,
-                    null)) {
+            try (OpenSSLClient client = createClient(
+                    certTuple, clientProtocol, clientCipherSuite,
+                    clientNamedGroup, clientSignatureScheme,
+                    isUseSessTicket,
+                    SESS_FILE, true)) {
                 client.connect("127.0.0.1", server.getPort());
-                context = client.context;
-                firstCreationTime = client.getSession().getCreationTime();
+                firstCreationTime = server.getSession().getCreationTime();
             }
 
-            try (JdkClient client = createClient(
-                    certTuple, clientProtocol,
-                    clientCipherSuite, clientNamedGroup,
-                    clientSignatureScheme,
-                    context)) {
+            try (OpenSSLClient client = createClient(
+                    certTuple, clientProtocol, clientCipherSuite,
+                    clientNamedGroup, clientSignatureScheme,
+                    isUseSessTicket,
+                    SESS_FILE, false)) {
                 client.connect("127.0.0.1", server.getPort());
 
-                long secondCreationTime = client.getSession().getCreationTime();
+                long secondCreationTime = server.getSession().getCreationTime();
                 Assertions.assertEquals(firstCreationTime, secondCreationTime);
             }
         } finally {
@@ -431,22 +427,37 @@ public class BabaSSLServerJdkClientTest {
         }
     }
 
-    private JdkClient createClient(
-            CertTuple certTuple, Protocol protocol,
+    private OpenSSLClient createClient(CertTuple certTuple, Protocol protocol,
             CipherSuite cipherSuite, NamedGroup namedGroup,
             SignatureScheme signatureScheme,
-            SSLContext context) throws Exception {
-        JdkClient.Builder builder = new JdkClient.Builder();
+            boolean isUseSessTicket,
+            Path sessFile, boolean saveSess) {
+        OpenSSLClient.Builder builder = new OpenSSLClient.Builder();
         builder.setCertTuple(certTuple);
         builder.setProtocols(protocol);
         builder.setCipherSuites(cipherSuite);
-//        builder.setNamedGroups(namedGroup);
-//        builder.setSignatureSchemes(signatureScheme);
-        builder.setMessage(
-                // An HTTP request asks to access the page.
-                String.format("GET /%s HTTP/1.1\r\n", PAGE_FILE));
-        builder.setReadResponse(true);
-        builder.setContext(context);
+        builder.setNamedGroups(namedGroup);
+        builder.setSignatureSchemes(signatureScheme);
+        builder.setMessage("Q");  // quit s_client
+        builder.setReadResponse(false);
+        builder.setUseSessTicket(isUseSessTicket);
+
+        if (sessFile != null) {
+            String sessFilePath = sessFile.toAbsolutePath().toString();
+            if (saveSess) {
+                builder.sessOut(sessFilePath);
+            } else {
+                builder.sessIn(sessFilePath);
+            }
+        }
+
         return builder.build();
+    }
+
+    private OpenSSLClient createClient(CertTuple certTuple, Protocol protocol,
+            CipherSuite cipherSuite, NamedGroup namedGroup,
+            SignatureScheme signatureScheme) {
+        return createClient(certTuple, protocol, cipherSuite, namedGroup,
+                signatureScheme, true, null, false);
     }
 }
